feat: renamed test folder
